Let the RESTART command reset any scdaemon pin prompts.
The gpg-agent sends the RESTART command to scdaemon when the client has
disconnected. This commit allows future connections to scdaemon use the
default pinentry prompts.

It is recommended to use the scdaemon LOCK command to prevent other
scdaemon clients from altering the prompts before the current client has
finished with the scdaemon pinentry since the set prompts are global to
the application.

* scd/command.c (cmd_restart): Free any allocated custom pinentry prompt
strings to let the application use its default.